+++ src/PEGParser.jl
@@ -1,29 +1,18 @@
 module PEGParser
-export @grammar
-export StandardCache, Node, transform, Grammar, Rule
-export no_action, or_default_action
-export ParserData, MatchRule
-export ?, list, parseGrammar, parseDefinition
 
-#using Compat
+export StandardCache, Node, transform, Grammar, Rule, MatchRule
+export no_action, liftchild
+export parseGrammar, parseDefinition
+
 import Base: show, parse, ==
-import Base: +, |, *, ^, >, -, !, integer, float # TODO: these should not even conflict with Base
-
-function no_action(rule, value, first, last, children)
-  return Node(rule.name, value, first, last, children, typeof(rule))
-end
-
-or_default_action(rule, value, first, last, children) = children[1]
-
-type MatchRule{T} end
 
 include("rules.jl")
 include("grammar.jl")
 include("comparison.jl")
-include("grammarparsing.jl")
+include("standardactions.jl")
 include("Node.jl")
 include("transform.jl")
 include("parse.jl")
-include("newgrammar.jl")
+include("grammarparsing.jl")
 
 end
